refactor: localStorage key값 상수화
